+++ core/src/main/kotlin/be.catvert.pc/utility/Rect.kt
@@ -16,7 +16,7 @@ class Rect(position: Point = Point(), size: Size = Size()) : CustomEditorImpl {
         }
     }
 
-    constructor(x: Int, y: Int, width: Int, height: Int): this(Point(x, y), Size(width, height))
+    constructor(x: Float, y: Float, width: Int, height: Int): this(Point(x, y), Size(width, height))
     constructor(rect: Rect): this(rect.x, rect.y, rect.width, rect.height)
 
     @JsonIgnore var position = position
@@ -33,12 +33,12 @@ class Rect(position: Point = Point(), size: Size = Size()) : CustomEditorImpl {
     @JsonIgnore val onPositionChange = Signal<Point>()
     @JsonIgnore val onSizeChange = Signal<Size>()
 
-    var x: Int
+    var x: Float
         get() = position.x
         set(value) {
             position = Point(value, y)
         }
-    var y: Int
+    var y: Float
         get() = position.y
         set(value) {
             position = Point(x, value)
